backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
python-multipart>=0.0.9
openpyxl>=3.1.0
email-validator>=2.0.0
redis>=5.0.0
rq>=1.15.0
bcrypt>=4.1.2
sentry-sdk[fastapi]>=2.19.0

